ref: resolver renamed to processor
