+++ backend/molgenis-emx2-sql/src/main/java/org/molgenis/emx2/sql/JavascriptContextBuilder.java
@@ -1,4 +1,4 @@
-package org.molgenis.emx2.sql.resolvers;
+package org.molgenis.emx2.sql;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
